Added new PlanOverviewSectionPageShared component to share across the versionedSection and customSection components
