+++ components/PlanOverviewSectionPageShared/__mocks__/versionedQuestionMock.ts
@@ -4,6 +4,7 @@ const versionedQuestionsMock = [
   {
     id: 1,
     questionText: 'What types of data will be produced during your project?',
+    questionType: 'BASE',
     displayOrder: 1,
     guidanceText: 'Guidance for question 1',
     requirementText: 'Requirement for question 1',
@@ -16,6 +17,7 @@ const versionedQuestionsMock = [
   {
     id: 2,
     questionText: 'What type of metadata will be collected?',
+    questionType: 'BASE',
     displayOrder: 2,
     guidanceText: 'Guidance for question 2',
     requirementText: 'Requirement for question 2',
@@ -28,6 +30,7 @@ const versionedQuestionsMock = [
   {
     id: 3,
     questionText: 'Will all data be converted to open source formats?',
+    questionType: 'BASE',
     displayOrder: 3,
     guidanceText: 'Guidance for question 3',
     requirementText: 'Requirement for question 3',
